Create schema directory and create new files for each schema
