+++ packages/webcomponents/src/components/business-forms/business-form-stepped/business-core-info/business-core-info-form-step.tsx
@@ -1,10 +1,10 @@
 import { Component, Host, h, Prop, State, Method, Event, EventEmitter } from '@stencil/core';
-import { BusinessStructureOptions, BusinessTypeOptions } from '../business-form-schema';
+import { BusinessStructureOptions, BusinessTypeOptions } from '../../schemas/business-form-schema';
 import { FormController } from '../../../form/form';
 import { PHONE_MASKS, TAX_ID_MASKS } from '../../../../utils/form-input-masks';
 import { BusinessFormServerErrors, CoreBusinessInfo, IBusiness, ICoreBusinessInfo } from '../../../../api/Business';
 import { Api, IApiResponse } from '../../../../api';
-import { coreInfoSchema } from '../../business-form/business-form-schema';
+import { businessCoreInfoSchema } from '../../schemas/business-core-info-schema';
 import { config } from '../../../../../config';
 import { parseCoreInfo } from '../helpers';
 import { flattenNestedObject } from '../../../../utils/utils';
@@ -85,7 +85,7 @@ export class BusinessCoreInfoFormStep {
     if (!this.authToken) console.error(missingAuthTokenMessage);
     if (!this.businessId) console.error(missingBusinessIdMessage);
 
-    this.formController = new FormController(coreInfoSchema);
+    this.formController = new FormController(businessCoreInfoSchema);
     this.api = Api(this.authToken, config.proxyApiOrigin);
     this.fetchData();
   }
