+++ packages/webcomponents/src/components/business-forms/business-form/business-generic-info/business-generic-info.tsx
@@ -3,7 +3,7 @@ import { Component, Host, h, Prop, State } from '@stencil/core';
 import {
   BusinessStructureOptions,
   BusinessTypeOptions,
-} from '../business-form-schema';
+} from '../../schemas/business-form-schema';
 import { FormController } from '../../../form/form';
 import { PHONE_MASKS, TAX_ID_MASKS } from '../../../../utils/form-input-masks';
 import { CoreBusinessInfo, ICoreBusinessInfo } from '../../../../api/Business';
