+++ packages/webcomponents/src/components/business-forms/schemas/business-address-schema.ts
@@ -1,0 +1,17 @@
+import { object, string } from 'yup';
+
+export const legalAddressSchema = object({
+  line1: string().required('Enter street address'),
+  line2: string().nullable(),
+  city: string().required('Enter city'),
+  state: string().required('Select state'),
+  postal_code: string().required('Enter postal code'),
+  country: string().required('Select country'),
+});
+
+export const identityAddressSchema = object({
+  line1: string().required('Enter street address'),
+  city: string().required('Enter city'),
+  state: string().required('Select state'),
+  postal_code: string().required('Enter postal code'),
+});
